outsourced shared function in Utils.tsx
